Day 1 done
Still no fix to the get all actors issue. Might have to put them in a different part of the level hidden from sight and then get all actors. Maybe this wil cause issues with optimization and ease of use but it's the only thing that will seem to work